Move FCE tests to node tests
